requirements.txt
torch>=1.8.0
torchvision>=0.9.0
transformers>=4.5.0
datasets>=1.6.0
matplotlib>=3.3.0
numpy>=1.19.0
pandas>=1.1.0
pyyaml>=5.4.0
seaborn>=0.11.0
tqdm>=4.60.0
scikit-learn>=0.24.0 